Item Lookup When Searched and Pagination
The user can search from homepage shop.php page as of this commit. When
the user searches, database will use the keywords provided to check if
they are tags in "item.tags". If there is a match, it will display that
item on the screen. Pagination is dynamic based on the number of
results. As of this commit, have not checked pagination for more than 2
items on page. Max 20 items per page. IMPORTANT: WHEN CLICKING IMAGE TO
GO TO ITEM DESCRIPTION PAGE THE PORT NUMBER IS CODED TO BE 81! MUST
CHANGE BASED ON MAMP SETTINGS!