Checkout - READ NOTE IN CAPS
When checking out, the following happens. If the shipping address is
provided, then the it will auto populate in the address section, if not,
the user must enter before checkout can occur. Once address is entered,
the STREET ADDRESS, CITY, AND STATE ARE STORED IN THE TRANSACTION TABLE,
I put this in there in the event that the user wants to change the
address to send to from the address on their account, they can do so and
we can still ship it while attributing the transaction to a particular
user.  When user hits "finish checking out"  the tables "transaction",
"transaction_items", and "user_transactions" are simultaneously updated
to reflect the additions. The cart is also cleared out for that
particular user in the "cart" table.

Currently the cart uses ajax technology to show checkout has completed.
Maybe I should redirect or refresh after delay?

Must remove "Checkout" tab on our pages as it is no longer necessary.
All checkout is done in the cart.