More Items/Fixed Pagination/Clicking Shop Fix
This commit contains more items added to the item table in order for
demonstration. Number of values shown per page can be changed with the
variable in shop.html called $rowsPerPage.  When clicking shop a
comprehensive list of all items in the item database are shown.
Currently pagination does not work for search via categories but that is
an easy fix I will put put shortly.